requirements.txt
aiodns==3.0.0
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-requests==0.1.3
aiokafka==0.7.1
aiologger==0.4.0rc1
aiomysql==0.0.21
aiopg==1.3.1
aioredis==2.0.0
aiorwlock==1.0.0
async-timeout==3.0.1
attrs==21.2.0
cchardet==2.1.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
coloredlogs==15.0.1
DBUtils==2.0.2
Faker==8.12.0
humanfriendly==9.2
idna==3.2
kafka-python==2.0.2
logutils==0.3.5
multidict==5.1.0
prettytable==2.1.0
psycopg2-binary==2.9.1
pycares==4.0.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==0.9.3
python-dateutil==2.8.2
rainbow==2.7.1
rainbow-logging-handler==2.2.2
redis==3.5.3
requests==2.26.0
selenium==3.141.0
six==1.16.0
text-unidecode==1.3
tornado==6.1
tqdm==4.62.1
typing-extensions==3.10.0.0
urllib3==1.26.6
watchdog==2.1.4
wcwidth==0.2.5
yarl==1.6.3
